project/requirements.txt
amqp==5.2.0
asgiref==3.7.2
autopep8==2.0.4
billiard==4.2.0
celery==5.3.6
Django==5.0.2
django-allauth==0.61.1
django-crispy-forms==2.1
django-filter==23.5
django-silk==5.1.0
djangorestframework==3.14.0
flower==2.0.1
gprof2dot==2022.7.29
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
kombu==5.3.5
oauthlib==3.2.2
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.8.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
tornado==6.4
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.3
pytest-factoryboy==2.8.1